Automate a complete site update
-- The new auto_update.py dynamic content generator uses images and
   json data prepared by an upstream program.  It accesses them from a
   directory given as argv[1].  The generator 1) stages the per
   jurisdiction .rst files; 2) translates the upstream json into json
   that is used by the Jinja2 engine to update them and other key .rst
   files; 3) stages the images for access during the sphinx build; and
   4) creates an html_context json file that allows injection of
   upstream content during the normal sphinx build template
   specialization phase.

-- Critical to using jinja to update the .rst files is the
   sphinx-jinja extension written by Pierre Tardy and offered under
   the MIT license via the sphinx-contrib mechanism.  This project
   used sphinx-jinja via addition of Tardy's jinja.py file to the
   config directory as a local extension, not via sphinx-contrib.

-- With everything properly staged and the sphinx-jinja extension in
   place, the normal sphinx "make html" command generates the website
   based upon the upstream data.  This is done using an updated
   version of the simple build_and_ship shell script which also
   creates a tarball of the html product, sends it to the web server,
   and unpacks it there.  FIXME: there are hard codes in that script
   unique to my setup.

-- Supporting this are changes to the conf.py configuration file which
   now read json files to set key configuration variables based on
   dynamic content.

-- The text in the Methodology section is expanded.

+++ build_and_ship.sh
@@ -2,7 +2,8 @@
 
 SCRIPT_DIR=$(dirname $(readlink -e $0))
 BUILD_DIR=$(readlink -e $SCRIPT_DIR/../covid_build)
-BASE_NAME=$(sed -ne "s|^    'html_path_prefix' : '/\(.*\)',|\1|p" website/conf.py)
+BASE_NAME=$(cat $SCRIPT_DIR/html_path_prefix.txt)
+BASE_NAME=${BASE_NAME#/}
 SSH_HOST=webserv@192.168.13.11
 TARBALL_DEST=/var/www/vhosts/www.dericnet.com
 
@@ -11,12 +12,18 @@ which pipenv
 VENV_DIR=$(pipenv --venv)
 
 pipenv run make clean
+#export SPHINXOPTS="-v -v"
 pipenv run make html
 
 cd $BUILD_DIR
 
 mv html $BASE_NAME
-tar czf $BASE_NAME.tar.gz $BASE_NAME
 
+# if there are any arguments, just build, don't ship
+if [ $# -gt 0 ]; then
+    exit 0
+fi
+
+tar czf $BASE_NAME.tar.gz $BASE_NAME
 scp $BASE_NAME.tar.gz $SSH_HOST:$TARBALL_DEST
 ssh $SSH_HOST "cd $TARBALL_DEST; rm -rf $BASE_NAME; tar xzf $BASE_NAME.tar.gz; rm $BASE_NAME.tar.gz"
